+++ app/%5Fpress/[site]/page.tsx
@@ -1,0 +1,6 @@
+import { createBlogIndex, createSiteStaticParams } from "barakopress";
+import { config } from "@/press.config";
+
+export default createBlogIndex(config);
+export const generateStaticParams = createSiteStaticParams();
+export const revalidate = 300;