+++ app/categories/[slug]/page.tsx
@@ -1,5 +1,0 @@
-import { createArchive } from "barakopress";
-import { config } from "@/press.config";
-
-export default createArchive(config, "category");
-export const revalidate = 300;