+++ app/page.tsx
@@ -1,5 +1,0 @@
-import { createBlogIndex } from "barakopress";
-import { config } from "@/press.config";
-
-export default createBlogIndex(config);
-export const revalidate = 300;